+++ test/testEditor.js
@@ -1,27 +1,27 @@
 let editorInstance;
 
-module.exports.initEditor = function(options, init)
-{
+module.exports.initEditor = function(options, init) {
     if(typeof options === 'string') {
         init = options;
         options = {};
     }
-    let $stage = $("#stage");
-    $stage.html('<textarea id="stageInput"></textarea>');
-    editorInstance = new window.prosemirror.MarkdownEditor($stage, options);
+
+    options = options || {edit : true};
+
+    editorInstance = new window.prosemirror.MarkdownEditor("#stage", options);
 
     editorInstance.init(init);
 
     return editorInstance;
-}
+};
 
 module.exports.toHtml = function() {
     return $("#stage .ProseMirror").html();
-}
+};
 
 module.exports.serialize = function() {
-    return  getEditor().serialize();
-}
+    return getEditor().serialize();
+};
 
 let getEditor = function(editor) {
     if(editor) {
@@ -33,7 +33,7 @@ let getEditor = function(editor) {
     }
 
     return module.exports.initEditor();
-}
+};
 
 module.exports.getEditor = getEditor;
 
@@ -59,13 +59,13 @@ module.exports.type = function(word, editor, startPos) {
 
     setSelection(nextPos, null, editor);
 
-}
+};
 
 let insertText = function(text, from, to, editor)
 {
     editor = getEditor(editor);
     editor.view.dispatch(editor.view.state.tr.insertText(text, from, to));
-}
+};
 
 module.exports.insertText = insertText;
 
@@ -78,15 +78,41 @@ let setSelection = function(start, end, editor) {
 
     let selection = new prosemirror.state.TextSelection(editor.view.state.doc.resolve(start), editor.view.state.doc.resolve(end));
     editor.view.dispatch(editor.view.state.tr.setSelection(selection));
-}
+};
 
 module.exports.setSelection = setSelection;
 
 let clickMenuItem = function(id) {
-    $('.ProseMirror-menu-'+id).trigger('mousedown');
+    $('.ProseMirror-menu-'+id)[0].dispatchEvent(new Event('mousedown', {bubbles: true}));
+};
+
+let clickDropdownMenuItem = function(id, subId, subId2) {
+    clickMenuItem(id);
+    clickMenuItem(subId);
+    if(subId2) {
+        clickMenuItem(subId2);
+    }
+    // Manually close
+    clickMenuItem(id);
+};
+
+let menuItemDisabled = function(id) {
+    return $('.ProseMirror-menu-'+id).is('.ProseMirror-menu-disabled');
+};
+
+let pressKey = function(key, code) {
+    $('.ProseMirror')[0].dispatchEvent(new KeyboardEvent('keydown', { key: key, code: code }));
 }
 
+let pressKeyArrowDown = function() {
+    pressKey('ArrowDown', 40);
+}
+
+module.exports.pressKey = pressKey;
+module.exports.pressKeyArrowDown = pressKeyArrowDown;
+module.exports.clickDropdownMenuItem = clickDropdownMenuItem;
 module.exports.clickMenuItem = clickMenuItem;
+module.exports.menuItemDisabled = menuItemDisabled;
 
 module.exports.simulateInputRule = function(word, editor) {
     editor = getEditor(editor);
@@ -98,4 +124,9 @@ module.exports.simulateInputRule = function(word, editor) {
     return editor.view.someProp("handleTextInput", function (f) {
         return f(editor.view, word.length, word.length, trigger);
     });
+};
+
+module.exports.selectSource = function(start, end, direction, editor) {
+    editor = getEditor(editor);
+    editor.context.$source[0].setSelectionRange(start, end, direction);
 }